pyproject.toml
[project]
name = "bpmn-documentation-extractor"
version = "0.1.0"
description = "Extract structured documentation from BPMN 2.0 files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "lxml>=5",
    "jinja2>=3.1",
    "pyyaml>=6",
]

[project.scripts]
bpmn-doc = "bpmn_doc.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bpmn_doc"]
